backend/requirements.txt
boto3==1.35.29
cloudinary==1.41.0
Flask==3.0.3
Flask_Cors==5.0.0
groq==0.11.0
langchain==0.3.1
langchain_chroma==0.1.4
langchain_core==0.3.6
langchain_openai==0.2.1
openai==1.50.2
protobuf==5.28.2
python-dotenv==1.0.1
Requests==2.32.3
faiss-cpu
